requirements.txt
affine==2.4.0
asgiref==3.6.0
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
Django==4.2.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
docopt==0.6.2
drf-yasg==1.21.5
Fiona==1.9.3
fonttools==4.39.4
#GDAL==3.2.2
GeoAlchemy2==0.13.2
geopandas==0.13.0
geoserver-rest==2.4.1
greenlet==2.0.2
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Js2Py==0.74
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
munch==3.0.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pipwin==0.5.2
psycopg2==2.9.6
pycurl==7.45.2
Pygments==2.15.1
pyjsparser==2.7.1
pyparsing==3.0.9
PyPrind==2.11.3
pyproj==3.5.0
pySmartDL==1.3.4
python-dateutil==2.8.2
pytz==2023.3
rasterio==1.3.6
requests==2.30.0
Rtree==1.0.1
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
snuggs==1.4.7
soupsieve==2.4.1
SQLAlchemy==2.0.13
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==2.0.2
xmltodict==0.13.0
